fix: parse float input expense amount

+++ src/utils/formaters.ts
@@ -1,7 +1,7 @@
 import moment from "moment"
 
 export const formatAmount = (value: string, typeCurrency = "BRL", locale = "pt-BR") => {
-    const valueNumber = parseInt(value)
+    const valueNumber = parseFloat(value)
 
     const valueFormatted = new Intl.NumberFormat(
         locale,
